requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3

# DICOM Processing
pydicom==2.4.4
Pillow==10.1.0
numpy==1.26.2

# OCR for PHI Removal
pytesseract==0.3.10

# Template Management
python-docx==1.1.0

# LLM Integration
openai==1.6.1
requests==2.31.0
httpx==0.25.2

# UI Framework
streamlit==1.29.0
streamlit-extras==0.3.6

# Utilities
loguru==0.7.2
tqdm==4.66.1

# Data Handling
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1